requirements.txt
Flask==2.1.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
pathspec==0.9.0
psycopg2-binary==2.9.3
python-dotenv==0.19.0
python-dateutil==2.8.2
pytz==2022.1
SQLAlchemy==1.4.39
sqlalchemy-views==0.3.1
tomli==2.0.1
ujson==5.4.0
Werkzeug==2.1.2